+++ angel_container/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add `ThrowingReflector`, which throws on all operations.
 * `EmptyReflector` uses `Object` instead of `dynamic` as its returned
 type, as the `dynamic` type is (apparently?) no longer a valid constant value.
+* `registerSingleton` now returns the provided `object`.
+* `registerFactory` and `registerLazySingleton` now return the provided function `f`.
 
 # 1.0.4
 * Slight patch to prevent annoying segfault.